refactor: Move simple typeguards to Observable.ts
- Since we unified how we were checking the observable input types, we no longer need shared modules for each typeguard.
- Most of the files touched here were using `isFunction`, etc.
