+++ packages/rxjs/src/internal/observable/fromEvent.ts
@@ -1,6 +1,4 @@
-import { Observable, Subscriber } from '../Observable.js';
-import { isArrayLike } from '../util/isArrayLike.js';
-import { isFunction } from '../util/isFunction.js';
+import { Observable, Subscriber, isArrayLike, isFunction } from '../Observable.js';
 import { mapOneOrManyArgs } from '../util/mapOneOrManyArgs.js';
 
 // These constants are used to create handler registry functions using array mapping below.
