+++ packages/rxjs/src/internal/operators/tap.ts
@@ -1,6 +1,5 @@
 import { MonoTypeOperatorFunction, Observer } from '../types.js';
-import { isFunction } from '../util/isFunction.js';
-import { Observable, operate } from '../Observable.js';
+import { Observable, operate, isFunction } from '../Observable.js';
 import { identity } from '../util/identity.js';
 
 /**
