+++ packages/rxjs/src/internal/scheduled/scheduleIterable.ts
@@ -1,7 +1,6 @@
-import { Observable } from '../Observable.js';
+import { Observable, isFunction } from '../Observable.js';
 import { SchedulerLike } from '../types.js';
 import { iterator as Symbol_iterator } from '../symbol/iterator.js';
-import { isFunction } from '../util/isFunction.js';
 import { executeSchedule } from '../util/executeSchedule.js';
 
 /**
