+++ packages/rxjs/src/internal/scheduled/scheduleReadableStreamLike.ts
@@ -1,7 +1,6 @@
 import { SchedulerLike, ReadableStreamLike } from '../types.js';
-import { Observable } from '../Observable.js';
+import { Observable, readableStreamLikeToAsyncGenerator } from '../Observable.js';
 import { scheduleAsyncIterable } from './scheduleAsyncIterable.js';
-import { readableStreamLikeToAsyncGenerator } from '../util/isReadableStreamLike.js';
 
 export function scheduleReadableStreamLike<T>(input: ReadableStreamLike<T>, scheduler: SchedulerLike): Observable<T> {
   return scheduleAsyncIterable(readableStreamLikeToAsyncGenerator(input), scheduler);
