--- packages/rxjs/src/internal/util/isArrayLike.ts
@@ -1,1 +1,0 @@
-export const isArrayLike = (<T>(x: any): x is ArrayLike<T> => x && typeof x.length === 'number' && typeof x !== 'function');
