--- packages/rxjs/src/internal/util/isFunction.ts
@@ -1,7 +1,0 @@
-/**
- * Returns true if the object is a function.
- * @param value The value to check
- */
-export function isFunction(value: any): value is (...args: any[]) => any {
-  return typeof value === 'function';
-}
