+++ packages/rxjs/src/internal/util/isScheduler.ts
@@ -1,5 +1,5 @@
 import { SchedulerLike } from '../types.js';
-import { isFunction } from './isFunction.js';
+import { isFunction } from '../Observable.js';
 
 export function isScheduler(value: any): value is SchedulerLike {
   return value && isFunction(value.schedule);
